LRCI-1030 Add SearchResult class to store all queries made against Spira
