+++ modules/test/jenkins-results-parser/src/main/java/com/liferay/jenkins/results/parser/spira/SearchResult.java
@@ -1,0 +1,108 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.jenkins.results.parser.spira;
+
+import java.util.Objects;
+
+import org.json.JSONObject;
+
+/**
+ * @author Michael Hashimoto
+ */
+public class SearchResult<T extends SpiraArtifact> {
+
+	public static class SearchParameter {
+
+		public SearchParameter(String name, Object value) {
+			_name = name;
+			_value = value;
+		}
+
+		@Override
+		public boolean equals(Object object) {
+			if (object instanceof SearchParameter) {
+				SearchParameter otherSearchParameter = (SearchParameter)object;
+
+				if (_name.equals(otherSearchParameter.getName()) &&
+					_value.equals(otherSearchParameter.getValue())) {
+
+					return true;
+				}
+
+				return false;
+			}
+
+			return super.equals(object);
+		}
+
+		public String getName() {
+			return _name;
+		}
+
+		public Object getValue() {
+			return _value;
+		}
+
+		@Override
+		public int hashCode() {
+			JSONObject filterJSONObject = toFilterJSONObject();
+
+			return filterJSONObject.hashCode();
+		}
+
+		public boolean matches(JSONObject jsonObject) {
+			if (!Objects.equals(getValue(), jsonObject.get(getName()))) {
+				return false;
+			}
+
+			return true;
+		}
+
+		public JSONObject toFilterJSONObject() {
+			JSONObject filterJSONObject = new JSONObject();
+
+			filterJSONObject.put("PropertyName", _name);
+
+			if (_value instanceof Integer) {
+				Integer intValue = (Integer)_value;
+
+				filterJSONObject.put("IntValue", intValue);
+			}
+			else if (_value instanceof String) {
+				String stringValue = (String)_value;
+
+				stringValue = stringValue.replaceAll("\\[", "[[]");
+
+				filterJSONObject.put("StringValue", stringValue);
+			}
+			else {
+				throw new RuntimeException("Invalid value type");
+			}
+
+			return filterJSONObject;
+		}
+
+		private final String _name;
+		private final Object _value;
+
+	}
+
+	protected SearchResult(SearchParameter... searchParameters) {
+		_searchParameters = searchParameters;
+	}
+
+	private final SearchParameter[] _searchParameters;
+
+}
