+++ modules/test/jenkins-results-parser/src/main/java/com/liferay/jenkins/results/parser/spira/SpiraTestSetFolder.java
@@ -151,7 +151,8 @@ public class SpiraTestSetFolder extends IndentLevelSpiraArtifact {
 	}
 
 	protected static List<SpiraTestSetFolder> getSpiraTestSetFolders(
-		SpiraProject spiraProject, SearchParameter... searchParameters) {
+		SpiraProject spiraProject,
+		SearchResult.SearchParameter... searchParameters) {
 
 		List<SpiraTestSetFolder> spiraTestSetFolders = new ArrayList<>();
